# openapi.json
{
  "components": {
    "schemas": {
      "AcquisitionRun": {
        "additionalProperties": true,
        "properties": {
          "config_yaml": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Yaml"
          },
          "experiment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experiment Id"
          },
          "extra": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "microscope_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Microscope Id"
          },
          "mm_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mm Version"
          },
          "pycroflow_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pycroflow Version"
          },
          "raw_data_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Data Path"
          },
          "raw_retained": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Retained"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "required": [
          "id"
        ],
        "title": "AcquisitionRun",
        "type": "object"
      },
...
      "Artifact": {
        "properties": {
          "analysis_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Run Id"
          },
          "checksum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checksum"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uri"
          }
        },
        "title": "Artifact",
        "type": "object"
      },
      "BulkIngest": {
        "description": "A batch of rows across tables, inserted in one transaction (backfill).\n\nTaxonomy is applied first so child paths resolve; the rest follow the FK\nchain. Any omitted list is skipped.",
        "properties": {
          "acquisition_run": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/AcquisitionRun"
            },
            "title": "Acquisition Run",
            "type": "array"
          },
          "analysis_run": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/AnalysisRun"
            },
            "title": "Analysis Run",
            "type": "array"
          },
          "artifact": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Artifact"
            },
            "title": "Artifact",
            "type": "array"
          },
          "environment": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Environment"
            },
            "title": "Environment",
            "type": "array"
          },
          "experiment": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Experiment"
            },
            "title": "Experiment",
            "type": "array"
          },
          "feedback": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Feedback"
            },
            "title": "Feedback",
            "type": "array"
          },
          "fluidics_round": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FluidicsRound"
            },
            "title": "Fluidics Round",
            "type": "array"
          },
          "fov": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Fov"
            },
            "title": "Fov",
            "type": "array"
          },
          "illumination": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Illumination"
            },
            "title": "Illumination",
            "type": "array"
          },
          "interpretation": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Interpretation"
            },
            "title": "Interpretation",
            "type": "array"
          },
          "metrics": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Metrics"
            },
            "title": "Metrics",
            "type": "array"
          },
          "qc": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Qc"
            },
            "title": "Qc",
            "type": "array"
          },
          "reagent_provenance": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ReagentProvenance"
            },
            "title": "Reagent Provenance",
            "type": "array"
          },
          "resource_usage": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ResourceUsage"
            },
            "title": "Resource Usage",
            "type": "array"
          },
          "sample_morphology": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SampleMorphology"
            },
            "title": "Sample Morphology",
            "type": "array"
          },
          "sample_tag": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SampleTag"
            },
            "title": "Sample Tag",
            "type": "array"
          },
          "sample_taxonomy": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SampleTaxonomy"
            },
            "title": "Sample Taxonomy",
            "type": "array"
          },
          "target_channel": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/TargetChannel"
            },
            "title": "Target Channel",
            "type": "array"
          }
        },
        "title": "BulkIngest",
        "type": "object"
      },
      "BulkResult": {
        "properties": {
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Counts",
            "type": "object"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "counts",
          "total"
        ],
        "title": "BulkResult",
        "type": "object"
      },
      "CohortItem": {
        "description": "One ranked cohort member (closest tree-distance first).",
        "properties": {
          "acquisition_run_id": {
            "title": "Acquisition Run Id",
            "type": "string"
          },
          "experiment_id": {
            "title": "Experiment Id",
            "type": "string"
          },
          "taxon_id": {
            "title": "Taxon Id",
            "type": "string"
          },
          "taxon_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taxon Name"
          },
          "tree_distance": {
            "title": "Tree Distance",
            "type": "integer"
          }
        },
        "required": [
          "acquisition_run_id",
          "experiment_id",
          "taxon_id",
          "tree_distance"
        ],
        "title": "CohortItem",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NodeDefaults": {
        "description": "The inherited cascade resolved for a taxonomy node.",
        "properties": {
          "defaults": {
            "additionalProperties": true,
            "default": {},
            "title": "Defaults",
            "type": "object"
          },
          "expected_metrics": {
            "additionalProperties": true,
            "default": {},
            "title": "Expected Metrics",
            "type": "object"
          },
          "qc_rules": {
            "additionalProperties": true,
            "default": {},
            "title": "Qc Rules",
            "type": "object"
          },
          "taxon_id": {
            "title": "Taxon Id",
            "type": "string"
          }
        },
        "required": [
          "taxon_id"
        ],
        "title": "NodeDefaults",
        "type": "object"
      },
...
      "SampleTag": {
        "properties": {
          "experiment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experiment Id"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag"
          }
        },
        "title": "SampleTag",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Append-only provenance & metrics database for the DNA-PAINT automation stack. Everything joins on run_id (acquisition_run.id).",
    "title": "picasso-registry",
    "version": "contract"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/bulk": {
      "post": {
        "description": "Insert many rows across tables in one transaction (backfill).",
        "operationId": "bulk_ingest_bulk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkIngest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bulk Ingest",
        "tags": [
          "ingest"
        ]
      }
    },
    "/cohort": {
      "get": {
        "description": "Acquisition runs ranked by sample-taxon tree distance, optionally\nconstrained on the other A2 descriptor axes (register C12).\n\nRanking (axis 1) is unchanged: exact-node matches first, then falling\nback *up* the tree; only runs under the same taxonomy **root** are\nconsidered and ``max_distance`` caps how far to generalize.\n\nOn top of that, axis 2 (``target`` / ``target_set`` names, or the\nclosed ``target_class``) and axis 3 (``modality`` / ``dimensionality``\n/ ``buffer``) are **independent optional filters** \u2014 each narrows the\ncohort when supplied. *How much* must match is comparison-dependent and\nleft to the caller (the recommender/agent decides which axes to\nconstrain for a given metric); the registry does not hard-code that\npolicy. With no axis args this is exactly the S0B-1 taxon-only cohort.",
        "operationId": "query_cohort_cohort_get",
        "parameters": [
          {
            "in": "query",
            "name": "taxon_id",
            "required": true,
            "schema": {
              "title": "Taxon Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "max_distance",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Distance"
            }
          },
          {
            "in": "query",
            "name": "modality",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "TIRF",
                    "HILO",
                    "spinning_disk",
                    "light_sheet"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Modality"
            }
          },
          {
            "in": "query",
            "name": "dimensionality",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "2D",
                    "3D"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dimensionality"
            }
          },
          {
            "in": "query",
            "name": "buffer",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Buffer"
            }
          },
          {
            "in": "query",
            "name": "target",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target"
            }
          },
          {
            "in": "query",
            "name": "target_set",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Set"
            }
          },
          {
            "in": "query",
            "name": "target_class",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "intracellular_protein",
                    "membrane_protein",
                    "glycan"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Class"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CohortItem"
                  },
                  "title": "Response Query Cohort Cohort Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Query Cohort",
        "tags": [
          "query"
        ]
      }
    },
...
    "/node_defaults": {
      "get": {
        "description": "Inherited defaults for a node (descendant overrides ancestor).",
        "operationId": "node_defaults_node_defaults_get",
        "parameters": [
          {
            "in": "query",
            "name": "taxon_id",
            "required": true,
            "schema": {
              "title": "Taxon Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeDefaults"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Node Defaults",
        "tags": [
          "query"
        ]
      }
    },
...
  }
}